requirements.txt
alembic==1.16.5
async-timeout==5.0.1
bidict==0.23.1
blinker==1.9.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.1.8
dotenv==0.9.9
Faker==37.12.0
Flask==3.1.2
Flask-Migrate==4.1.0
Flask-SocketIO==5.6.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
h11==0.16.0
idna==3.11
importlib_metadata==8.7.1
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
netifaces==0.10.6
psycopg2-binary==2.9.11
python-dotenv==1.2.1
python-engineio==4.13.0
python-socketio==5.16.0
redis==7.0.1
requests==2.32.5
simple-websocket==1.1.0
spotipy==2.25.2
SQLAlchemy==2.0.46
tomli==2.4.0
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
Werkzeug==3.1.5
wsproto==1.2.0
zipp==3.23.0
